--- app/src/main/java/com/tpc/digigate/ui/components/AppTextField.kt
@@ -1,61 +1,0 @@
-package com.tpc.digigate.ui.components
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-@Composable
-fun AppTextField(
-    label: String,
-    value: String,
-    onValueChange: (String) -> Unit,
-    modifier: Modifier = Modifier,
-    placeholder: String = "",
-    errorMessage: String? = null,
-    isSingleLine: Boolean = true,
-    keyboardType: KeyboardType = KeyboardType.Text,
-    imeAction: ImeAction = ImeAction.Done,
-    onImeAction: () -> Unit = {},
-    leadingIcon: (@Composable (() -> Unit))? = null,
-    trailingIcon: (@Composable (() -> Unit))? = null
-){
-
-
-    var email by rememberSaveable {
-        mutableStateOf("")
-    }
-    Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center){
-
-        OutlinedTextField(modifier = Modifier.fillMaxWidth().padding(16.dp),shape = RoundedCornerShape(9.dp)
-            ,value = email, onValueChange = {email = it}, label = {
-                Text(text="Email")
-
-            },
-                    singleLine = true
-        )
-
-
-    }
-
-
-
-}
-@Preview(showBackground = true, showSystemUi = true)
-@Composable
-fun Preview(){
-    AppTextField(label = "Email", value = "", onValueChange = {})
-}
-
